Refactoring all classes from org.wildfly.loaders to org.wildfly.loaders.deployment package
