+++ loaders/src/main/java/org/wildfly/loaders/deployment/ResourceLoaders.java
@@ -20,11 +20,11 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.loaders;
+package org.wildfly.loaders.deployment;
 
-import static org.wildfly.loaders.Utils.explodeArchive;
-import static org.wildfly.loaders.Utils.getResourceName;
-import static org.wildfly.loaders.Utils.normalizePath;
+import static org.wildfly.loaders.deployment.Utils.explodeArchive;
+import static org.wildfly.loaders.deployment.Utils.getResourceName;
+import static org.wildfly.loaders.deployment.Utils.normalizePath;
 
 import java.io.File;
 import java.io.IOException;
