+++ loaders/src/test/java/org/wildfly/loaders/deployment/application.xml
