+++ loaders/src/test/java/org/wildfly/loaders/deployment/web.xml
@@ -32,7 +32,7 @@
 
     <servlet>
         <servlet-name>GreetingServlet</servlet-name>
-        <servlet-class>org.wildfly.loaders.GreetingServlet</servlet-class>
+        <servlet-class>org.wildfly.loaders.deployment.GreetingServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
