+++ server/src/main/java/org/jboss/as/server/deployment/annotation/ResourceRootIndexer.java
@@ -38,7 +38,7 @@ import org.jboss.jandex.Index;
 import org.jboss.jandex.IndexReader;
 import org.jboss.jandex.Indexer;
 import org.jboss.modules.Resource;
-import org.wildfly.loaders.ResourceLoader;
+import org.wildfly.loaders.deployment.ResourceLoader;
 
 /**
  * Utility class for indexing a resource root
