+++ server/src/test/java/org/jboss/as/server/deployment/module/ClassPathUtilsTestCase.java
@@ -30,7 +30,7 @@ import org.jboss.modules.ClassSpec;
 import org.jboss.modules.PackageSpec;
 import org.jboss.modules.Resource;
 import org.junit.Test;
-import org.wildfly.loaders.ResourceLoader;
+import org.wildfly.loaders.deployment.ResourceLoader;
 
 import java.io.File;
 import java.io.IOException;
